validate: Split application flow tests to application_test.go
Move all TestValidateApplication* tests and application-specific mock
instances (applicationMock, gatherS3Failed, etc.) from command_test.go
to application_test.go. This leaves command_test.go with only the
common test (TestValidatedDeleted) and clusters flow tests.
